moved LED unit test code to test_led_driver.c/.h
