+++ test/Test_List.md
@@ -1,0 +1,9 @@
+## Test List for Demonstration Widget
+
+The following is the list of unit tests for each of the peripherals on the microcontroller and external device drivers.
+
+#LED DRIVER
+- Initialize LEDs
+- De-Initialize LEDs 
+- A single LED can be turned on/off
+- Check each LED state
