+++ test/test_main.c
@@ -2,31 +2,11 @@
 #include <unity.h>
 
 void setUp(void) {
-    LED_GPIO_CLK_ENABLE();
-    GPIO_InitTypeDef GPIO_InitStruct;
-    GPIO_InitStruct.Pin = LED_PIN;
-    GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-    GPIO_InitStruct.Pull = GPIO_PULLUP;
-    GPIO_InitStruct.Speed = GPIO_SPEED_HIGH;
-    HAL_GPIO_Init(LED_GPIO_PORT, &GPIO_InitStruct);
+    LED_Init();
 }
 
 void tearDown(void) {
-    HAL_GPIO_DeInit(LED_GPIO_PORT, LED_PIN);
-}
-
-void test_led_builtin_pin_number(void) {
-    TEST_ASSERT_EQUAL(GPIO_PIN_5, LED_PIN);
-}
-
-void test_led_state_high(void) {
-    HAL_GPIO_WritePin(LED_GPIO_PORT, LED_PIN, GPIO_PIN_SET);
-    TEST_ASSERT_EQUAL(GPIO_PIN_SET, HAL_GPIO_ReadPin(LED_GPIO_PORT, LED_PIN));
-}
-
-void test_led_state_low(void) {
-    HAL_GPIO_WritePin(LED_GPIO_PORT, LED_PIN, GPIO_PIN_RESET);
-    TEST_ASSERT_EQUAL(GPIO_PIN_RESET, HAL_GPIO_ReadPin(LED_GPIO_PORT, LED_PIN));
+    LED_DeInit();
 }
 
 int main() {
